requirements.txt
findspark
pyspark
python-dateutil
pathlib
kafka-python
confluent_kafka
tornado